+++ src/components/layout/header/side-menu/SideMenu.tsx
@@ -1,14 +1,5 @@
-import {
-  BookOpenIcon,
-  MegaphoneIcon,
-  UserGroupIcon,
-  UserIcon,
-} from '@heroicons/react/24/outline'
-
-import { Button, SideMenuFooter, SideMenuHeader, Text } from '@components'
+import { SideMenuFooter, SideMenuHeader, SideMenuNav, Text } from '@components'
 import { cn } from '@utils'
-import { HEADER_MOBILE_NAV_LISTS } from '@constants'
-import { Link, useLocation } from 'react-router'
 
 interface HeaderSideMenuProps {
   isLoggedin: boolean
@@ -16,20 +7,11 @@ interface HeaderSideMenuProps {
   onClose: () => void
 }
 
-const NAV_ICONS = [
-  { name: '강의', icon: <BookOpenIcon width={18} /> },
-  { name: '그룹', icon: <UserGroupIcon width={18} /> },
-  { name: '공고', icon: <MegaphoneIcon width={18} /> },
-  { name: '내정보', icon: <UserIcon width={18} /> },
-]
-
 export default function HeaderSideMenu({
   isLoggedin,
   isOpen,
   onClose,
 }: HeaderSideMenuProps) {
-  const { pathname } = useLocation()
-
   return (
     <div className="fixed inset-0 z-50">
       {/* Overlay */}
@@ -53,26 +35,7 @@ export default function HeaderSideMenu({
             >
               메뉴
             </Text>
-            <nav>
-              <ul className="flex flex-col gap-2">
-                {HEADER_MOBILE_NAV_LISTS.map((el) => (
-                  <Link to={el.path}>
-                    <Button
-                      variant="ghost"
-                      size="large"
-                      className={cn(
-                        'hover:bg-primary-50 w-full gap-3 px-3',
-                        pathname.startsWith(el.path) &&
-                          'text-primary-600 bg-primary-50 font-medium'
-                      )}
-                    >
-                      {NAV_ICONS.find((icon) => el.name === icon.name)?.icon}
-                      {el.name}
-                    </Button>
-                  </Link>
-                ))}
-              </ul>
-            </nav>
+            <SideMenuNav />
           </div>
         </div>
