+++ src/components/layout/header/side-menu/index.ts
@@ -1,3 +1,4 @@
 export { default as SideMenu } from '@components/layout/header/side-menu/SideMenu'
 export { default as SideMenuHeader } from '@components/layout/header/side-menu/SideMenuHeader'
 export { default as SideMenuFooter } from '@components/layout/header/side-menu/SideMenuFooter'
+export { default as SideMenuNav } from '@components/layout/header/side-menu/SideMenuNav'
